[fentry] Hopefully fix sendpage on newer kernels
