+++ pkg/network/tracer/connection/kprobe/config.go
@@ -19,26 +19,9 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/network/config"
 	netebpf "github.com/DataDog/datadog-agent/pkg/network/ebpf"
 	"github.com/DataDog/datadog-agent/pkg/network/ebpf/probes"
+	"github.com/DataDog/datadog-agent/pkg/network/tracer/connection/util"
 	"github.com/DataDog/datadog-agent/pkg/util/kernel"
-	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
-
-// HasTCPSendPage checks if the kernel has the tcp_sendpage function.
-// After kernel 6.5.0, tcp_sendpage and udp_sendpage are removed.
-// We used to only check for kv < 6.5.0 here - however, OpenSUSE 15.6 backported
-// this change into 6.4.0 to pick up a CVE so the version number is not reliable.
-// Instead, we directly check if the function exists.
-func HasTCPSendPage(kv kernel.Version) bool {
-	missing, err := ebpf.VerifyKernelFuncs("tcp_sendpage")
-	if err == nil {
-		return len(missing) == 0
-	}
-
-	log.Debugf("unable to determine whether tcp_sendpage exists, using kernel version instead: %s", err)
-
-	kv650 := kernel.VersionCode(6, 5, 0)
-	return kv < kv650
-}
 
 func enableProbe(enabled map[probes.ProbeFuncName]struct{}, name probes.ProbeFuncName) {
 	enabled[name] = struct{}{}
@@ -74,7 +57,7 @@ func enabledProbes(c *config.Config, runtimeTracer, coreTracer bool) (map[probes
 	}
 	// else: Kernel < 4.15 - keep kprobe fallback (no multiple tracepoint attachment)
 
-	hasSendPage := HasTCPSendPage(kv)
+	hasSendPage := util.HasTCPSendPage(kv)
 
 	if c.CollectTCPv4Conns || c.CollectTCPv6Conns {
 		if ClassificationSupported(c) {
